requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==25.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
djangorestframework==3.15.2
importlib_metadata==8.6.1
Markdown==3.7
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.21.0
